+++ PersistencePlugin/src/main/java/com/elmakers/mine/bukkit/plugins/persistence/PersistencePlugin.java
@@ -8,7 +8,6 @@ import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.event.Event.Priority;
 import org.bukkit.event.Event.Type;
-import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.PluginDescriptionFile;
 import org.bukkit.plugin.PluginLoader;
 import org.bukkit.plugin.PluginManager;
@@ -16,8 +15,6 @@ import org.bukkit.plugin.java.JavaPlugin;
 
 import com.elmakers.mine.bukkit.plugins.persistence.core.PersistenceCommands;
 import com.elmakers.mine.bukkit.plugins.persistence.core.PersistenceListener;
-
-import com.nijikokun.bukkit.Permissions.Permissions;
 
 /** 
  * The JavaPlugin interface for Persistence- binds Persistence to Bukkit.
@@ -153,19 +150,6 @@ public class PersistencePlugin extends JavaPlugin
 		
 		pm.registerEvent(Type.PLAYER_QUIT, listener, Priority.Normal, this);
 		pm.registerEvent(Type.PLAYER_JOIN, listener, Priority.Normal, this);
-		
-		bindToPermissions();
-	}
-	
-	public void bindToPermissions()
-	{
-		Plugin checkForPermissions = this.getServer().getPluginManager().getPlugin("Permissions");
-	    if (checkForPermissions != null) 
-	    {
-	    	persistence.setPermissions((Permissions)checkForPermissions);
-	    	log.info("Persistence: Found Permissions, using it for permissions.");
-	    } 
-
 	}
 	
 	/*