+++ PersistencePlugin/src/main/java/com/elmakers/mine/bukkit/plugins/persistence/dao/PluginCommand.java
@@ -219,13 +219,6 @@ public class PluginCommand implements Comparable<PluginCommand>
 		switch (permissionType)
 		{
 			case ALLOW_ALL: return true;
-			case OPS_DEFAULT:
-				if (permissionNode != null && permissionNode.length() > 0)
-				{
-					// isOp used as default, Permissions can override.
-					return persistence.hasPermission(player, permissionNode, player.isOp());
-				}
-				break;
 			case OPS_ONLY:
 				if (player == null) return false;
 				return player.isOp();
@@ -238,7 +231,7 @@ public class PluginCommand implements Comparable<PluginCommand>
 			case ADMINS_ONLY:
 				if (permissionNode != null && permissionNode.length() > 0)
 				{
-					return persistence.hasPermission(player, permissionNode, false);
+					return persistence.hasPermission(player, permissionNode);
 				}
 				break;
 		}
@@ -306,6 +299,10 @@ public class PluginCommand implements Comparable<PluginCommand>
 					sender.sendMessage(currentIndent + " ex: " + getPath() + " " + exampleUse);
 				}
 			}
+		}
+		else
+		{
+			showSubCommands = true;
 		}
 		
 		if (showSubCommands && children != null)